+++ src/lambdas/api/app.js
@@ -6,7 +6,7 @@ const createError = require('http-errors')
 const express = require('express')
 const logger = require('morgan')
 const path = require('path')
-const satlib = require('../../lib')
+const es = require('../../lib/es')
 const api = require('../../lib/api')
 const { readYaml } = require('../../lib/fs')
 const { addEndpoint } = require('./middleware/add-endpoint')
@@ -30,7 +30,7 @@ app.use(addEndpoint)
 
 app.get('/', async (req, res, next) => {
   try {
-    res.json(await api.getCatalog(satlib.es, req.endpoint))
+    res.json(await api.getCatalog(txnEnabled, es, req.endpoint))
   } catch (error) {
     next(error)
   }
@@ -48,7 +48,7 @@ app.get('/api', async (_req, res, next) => {
 
 app.get('/conformance', async (_req, res, next) => {
   try {
-    res.json(await api.getConformance())
+    res.json(await api.getConformance(txnEnabled))
   } catch (error) {
     next(error)
   }
@@ -57,7 +57,7 @@ app.get('/conformance', async (_req, res, next) => {
 app.get('/search', async (req, res, next) => {
   try {
     res.type('application/geo+json')
-    res.json(await api.searchItems(null, req.query, satlib.es, req.endpoint, 'GET'))
+    res.json(await api.searchItems(null, req.query, es, req.endpoint, 'GET'))
   } catch (error) {
     next(error)
   }
@@ -66,7 +66,7 @@ app.get('/search', async (req, res, next) => {
 app.post('/search', async (req, res, next) => {
   try {
     res.type('application/geo+json')
-    res.json(await api.searchItems(null, req.body, satlib.es, req.endpoint, 'POST'))
+    res.json(await api.searchItems(null, req.body, es, req.endpoint, 'POST'))
   } catch (error) {
     next(error)
   }
@@ -74,7 +74,7 @@ app.post('/search', async (req, res, next) => {
 
 app.get('/collections', async (req, res, next) => {
   try {
-    res.json(await api.getCollections(satlib.es, req.endpoint))
+    res.json(await api.getCollections(es, req.endpoint))
   } catch (error) {
     next(error)
   }
@@ -83,7 +83,7 @@ app.get('/collections', async (req, res, next) => {
 app.get('/collections/:collectionId', async (req, res, next) => {
   const { collectionId } = req.params
   try {
-    const response = await api.getCollection(collectionId, satlib.es, req.endpoint)
+    const response = await api.getCollection(collectionId, es, req.endpoint)
 
     if (response instanceof Error) next(createError(404))
     else res.json(response)
@@ -95,14 +95,14 @@ app.get('/collections/:collectionId', async (req, res, next) => {
 app.get('/collections/:collectionId/items', async (req, res, next) => {
   const { collectionId } = req.params
   try {
-    const response = await api.getCollection(collectionId, satlib.es, req.endpoint)
+    const response = await api.getCollection(collectionId, es, req.endpoint)
 
     if (response instanceof Error) next(createError(404))
     else {
       const items = await api.searchItems(
         collectionId,
         req.query,
-        satlib.es,
+        es,
         req.endpoint,
         'GET'
       )
@@ -122,11 +122,11 @@ app.post('/collections/:collectionId/items', async (req, res, next) => {
     if (req.body.collection && req.body.collection !== collectionId) {
       next(createError(400, 'Collection resource URI must match collection in body'))
     } else {
-      const collectionRes = await api.getCollection(collectionId, satlib.es, req.endpoint)
+      const collectionRes = await api.getCollection(collectionId, es, req.endpoint)
       if (collectionRes instanceof Error) next(createError(404))
       try {
         req.body.collection = collectionId
-        await api.createItem(req.body, satlib.es)
+        await api.createItem(req.body, es)
         res.location(`${req.endpoint}/collections/${collectionId}/items/${itemId}`)
         res.sendStatus(201)
       } catch (error) {
@@ -151,7 +151,7 @@ app.get('/collections/:collectionId/items/:itemId', async (req, res, next) => {
     const response = await api.getItem(
       collectionId,
       itemId,
-      satlib.es,
+      es,
       req.endpoint
     )
 
@@ -179,13 +179,13 @@ app.put('/collections/:collectionId/items/:itemId', async (req, res, next) => {
     } else if (req.body.id && req.body.id !== itemId) {
       next(createError(400, 'Item ID in resource URI must match id in body'))
     } else {
-      const collectionRes = await api.getCollection(collectionId, satlib.es, req.endpoint)
+      const collectionRes = await api.getCollection(collectionId, es, req.endpoint)
       if (collectionRes instanceof Error) next(createError(404))
 
       req.body.collection = collectionId
       req.body.id = itemId
       try {
-        await api.updateItem(req.body, satlib.es)
+        await api.updateItem(req.body, es)
         res.sendStatus(204)
       } catch (error) {
         if (error instanceof Error
@@ -211,16 +211,16 @@ app.patch('/collections/:collectionId/items/:itemId', async (req, res, next) => 
     } else if (req.body.id && req.body.id !== itemId) {
       next(createError(400, 'Item ID in resource URI must match id in body'))
     } else {
-      const collectionRes = await api.getCollection(collectionId, satlib.es, req.endpoint)
+      const collectionRes = await api.getCollection(collectionId, es, req.endpoint)
       if (collectionRes instanceof Error) next(createError(404))
-      const itemRes = await api.getItem(collectionId, itemId, satlib.es, req.endpoint)
+      const itemRes = await api.getItem(collectionId, itemId, es, req.endpoint)
       if (itemRes instanceof Error) next(createError(404))
 
       else {
         try {
           //const item =
           await api.partialUpdateItem(
-            collectionId, itemId, req.body, satlib.es, req.endpoint
+            collectionId, itemId, req.body, es, req.endpoint
           )
           // res.type('application/geo+json')
           // res.json(item)
@@ -239,7 +239,7 @@ app.delete('/collections/:collectionId/items/:itemId', async (req, res, next) =>
   if (txnEnabled) {
     const { collectionId, itemId } = req.params
     try {
-      const response = await api.deleteItem(collectionId, itemId, satlib.es)
+      const response = await api.deleteItem(collectionId, itemId, es)
       if (response instanceof Error) next(createError(500))
       else {
         res.sendStatus(204)